El detalle de producto obtiene los productos de desayuno seleccionados
