+++ my-app/src/components/Breakfast.js
@@ -1,10 +1,8 @@
 import React, { useState, useEffect } from 'react';
 // import { collection, onSnapshot } from 'firebase/firestore'; // Importar Firestore
 import { db } from '../utils/firebaseConfig';
-import { doc, getDoc } from "firebase/firestore";
-import {PurchaseOrder} from './PurchaseOrder';
 
-function Breakfast() {
+function Breakfast(props) {
   // Data de desayuno
   const [breakfast, setBreakfast] = useState([]);
   // console.log(breakfast);
@@ -30,28 +28,6 @@ function Breakfast() {
   }, [])
 
   // -----Obtener data de producto seleccionado por mesero (PRUEBA) -----//
-  const [dataProducts, setDataProducts] = useState([]);
-  // console.log(dataProducts, 'final')
-
-  const addProduct = async (id) => {
-
-    // Acceder a la data del producto seleccionado
-    const breakfastRef = doc(db, "Desayunos", id);
-    
-    // Traer la data
-    const docSnap = await getDoc(breakfastRef);
-    const docData = docSnap.data();
-
-      const dataObj = {
-        id: id,
-        name: docData.name,
-        price: docData.price,
-      }
-      // console.log(dataObj)
-      setDataProducts(dataObj)
-
-
-  }
 
   return (
     <>
@@ -69,12 +45,11 @@ function Breakfast() {
             <h3>{product.name}</h3>
           </div>
           {/* <button type="button">AGREGAR</button> */}
-          <button type="button" onClick={()=> addProduct(product.id)}>AGREGAR</button>
+          <button type="button" onClick={()=> props.addProduct(product.id)}>AGREGAR</button>
         </div>
       </div>
     ))}
     </main>
-    <PurchaseOrder style={{ display: 'hidden' }} dataProducts={dataProducts}/>
     </>
   )
 }
